requirements (2).txt
nose>=1.3.7
python-dateutil>=1.5
urllib3>=1.18.1
pymongo>=3.4.0
websocket-client>=0.37.0
arrow>=0.10.0
python_jsonschema_objects>=0.2.1
httplib2>=0.10.3
appdirs>=1.4.3
configobj>=5.0.6
fauxfactory>=2.0.9
geopy>=1.11.0
googlemaps==2.5.1
oauthlib>=2.0.1
packaging>=16.8
polyline>=1.3.2
pyparsing>=2.2.0
requests>=2.11.1
requests-oauthlib>=0.8.0
requests-toolbelt>=0.7.1
six>=1.10.0
unittest-data-provider>=1.0.1
validictory>=1.1.0
pytest>=3.0.5
names>=0.3.0
locustio>=0.7
pyzmq>=16.0.2
concurrencytest>=0.1.2
pytest-xdist>=1.22.0
beautifulsoup4>=4.6.0